+++ src/store/modules/index.js
@@ -1,5 +1,7 @@
 import Page from '@/store/modules/page'
+import User from '@/store/modules/user'
 
 export default {
   Page,
+  User,
 }